pyproject.toml
[tool.poetry]
name = "flix-api"
version = "0.1.0"
description = ""
authors = ["Kauanrodrigues01m <kauanrl09@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.5"
python-decouple = "^3.8"
djangorestframework = "^3.15.2"
django-filter = "^24.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
